+++ src/lib/battery/battery.cpp
@@ -42,6 +42,7 @@
 #include "battery.h"
 #include <mathlib/mathlib.h>
 #include <cstring>
+#include <px4_defines.h>
 
 Battery::Battery() :
 	ModuleParams(nullptr),
