+++ src/lib/conversion/rotation.cpp
@@ -37,6 +37,7 @@
  * Vector rotation library
  */
 
+#include <px4_defines.h>
 #include "math.h"
 #include "rotation.h"
