+++ src/lib/terrain_estimation/terrain_estimator.cpp
@@ -38,6 +38,7 @@
 
 #include "terrain_estimator.h"
 #include <lib/ecl/geo/geo.h>
+#include <px4_defines.h>
 
 #define DISTANCE_TIMEOUT 100000		// time in usec after which laser is considered dead
